requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
dj-database-url==2.2.0
Django==5.0.6
django-admin-search==0.3.15
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-crispy-forms==2.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-otp==1.5.0
django-recaptcha==4.0.0
et-xmlfile==1.1.0
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
Pillow==9.5.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
razorpay==1.4.2
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tinify==1.6.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
XlsxWriter==3.2.0
